+++ CHANGELOG.md
@@ -51,3 +51,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added parameter group selection
 - changed Jenkins automatic version incrementing to tag with current version number on release then commit new version number
 - Added support for queries containing more than 100 parameter codes
+- Patched schema construction to allow repeated parameters from same site
+- Changed tests to use mocked files instead of fetched values